+++ notebooks/py_scripts/04-automate-optional.py
@@ -1,14 +1,13 @@
 # ---
 # jupyter:
 #   jupytext:
-#     formats: ipynb,py_scripts//py
 #     text_representation:
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.14.0
+#       jupytext_version: 1.13.7
 #   kernelspec:
-#     display_name: Python 3.10.4 64-bit ('python3p10')
+#     display_name: Python 3.9.12 ('ele')
 #     language: python
 #     name: python3
 # ---
@@ -20,8 +19,7 @@
 # -
 
 import os
-
-os.chdir("..")
+if os.path.basename(os.getcwd()) == "notebooks": os.chdir("..")
 
 import numpy as np
 from workflow_array_ephys.pipeline import lab, subject, session, probe, ephys