+++ notebooks/py_scripts/05-explore.py
@@ -1,14 +1,13 @@
 # ---
 # jupyter:
 #   jupytext:
-#     formats: ipynb,py_scripts//py
 #     text_representation:
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.14.0
+#       jupytext_version: 1.13.7
 #   kernelspec:
-#     display_name: Python 3.10.4 64-bit ('python3p10')
+#     display_name: Python 3.9.12 ('ele')
 #     language: python
 #     name: python3
 # ---
@@ -18,8 +17,7 @@
 # This notebook will describe the steps for interacting with the data ingested into `workflow-array-ephys`.
 
 import os
-
-os.chdir("..")
+if os.path.basename(os.getcwd()) == "notebooks": os.chdir("..")
 
 # +
 import datajoint as dj