+++ notebooks/py_scripts/09-NWB-export.py
@@ -1,14 +1,13 @@
 # ---
 # jupyter:
 #   jupytext:
-#     formats: ipynb,py_scripts//py
 #     text_representation:
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
 #       jupytext_version: 1.14.0
 #   kernelspec:
-#     display_name: Python 3.10.4 64-bit ('python3p10')
+#     display_name: Python 3.9.12 ('ele')
 #     language: python
 #     name: python3
 # ---
@@ -22,39 +21,40 @@
 # First, let's change directories to find the `dj_local_conf` file.
 
 import os
-
 # change to the upper level folder to detect dj_local_conf.json
-if os.path.basename(os.getcwd()) == "notebooks":
-    os.chdir("..")
-assert os.path.basename(os.getcwd()) == "workflow-array-ephys", (
-    "Please move to the " + "workflow directory"
-)
+if os.path.basename(os.getcwd())=='notebooks': os.chdir('..')
 
 # We'll be working with long tables, so we'll make visualization easier with a limit
 import datajoint as dj
+dj.config['display.limit']=10
 
-dj.config["display.limit"] = 10
+# CodeBook users should also set a couple additional config parameters.
 
-# If you haven't already populated the `lab`, `subject`, `session`, `probe`, and `ephys` schemas, please do so now with [04-automate](./04-automate-optional.ipynb). Note: exporting `ephys` data is currently only supported on the `no_curation` schema.
+username_as_prefix = dj.config["database.user"] + "_"
+if not dj.config['custom']:
+    dj.config['custom'] = {}
+dj.config['custom'].update({
+        "ephys_mode": "no-curation",
+        "database.prefix": username_as_prefix,
+        })
+
+# If you haven't already populated the `lab`, `subject`, `session`, `probe`, and `ephys` schemas, please do so now with [04-automate](./04-automate-optional.ipynb). Note: exporting `ephys` data is currently only supported on the `no_curation` schema. 
 
 from workflow_array_ephys.pipeline import lab, subject, session, probe, ephys
-from workflow_array_ephys.export import (
-    element_lab_to_nwb_dict,
-    subject_to_nwb,
-    session_to_nwb,
-    ecephys_session_to_nwb,
-    write_nwb,
-)
+from workflow_array_ephys.export import (element_lab_to_nwb_dict, subject_to_nwb, 
+                                         session_to_nwb, ecephys_session_to_nwb, 
+                                         write_nwb)
 from element_interface.dandi import upload_to_dandi
 
 # ## Export to NWB
 #
 # We'll use the following keys to demonstrate export functions.
 
-lab_key = {"lab": "LabA"}
-protocol_key = {"protocol": "ProtA"}
-project_key = {"project": "ProjA"}
-session_key = {"subject": "subject5", "session_datetime": "2018-07-03 20:32:28"}
+lab_key={"lab": "LabA"}
+protocol_key={"protocol": "ProtA"}
+project_key={"project": "ProjA"}
+session_key={"subject": "subject5",
+             "session_datetime": "2018-07-03 20:32:28"}
 
 # ### Upstream Elements
 #
@@ -67,13 +67,12 @@ session_key = {"subject": "subject5", "session_datetime": "2018-07-03 20:32:28"}
 # Note: `pynwb` will display a warning regarding timezone information - datetime fields are assumed to be in local time, and will be converted to UTC.
 #
 
-print("Lab:\n")
-element_lab_to_nwb_dict(
-    lab_key=lab_key, protocol_key=protocol_key, project_key=project_key
-)
-print("\nAnimal:\n")
+print('Lab:\n')
+element_lab_to_nwb_dict(lab_key=lab_key, protocol_key=protocol_key, 
+                        project_key=project_key)
+print('\nAnimal:\n')
 subject_to_nwb(session_key=session_key)
-print("\nSession:\n")
+print('\nSession:\n')
 session_to_nwb(session_key=session_key)
 
 # ### Element Array Electrophysiology
@@ -83,17 +82,15 @@ session_to_nwb(session_key=session_key)
 
 help(ecephys_session_to_nwb)
 
-nwbfile = ecephys_session_to_nwb(
-    session_key=session_key,
-    raw=True,
-    spikes=True,
-    lfp="dj",
-    end_frame=100,
-    lab_key=lab_key,
-    project_key=project_key,
-    protocol_key=protocol_key,
-    nwbfile_kwargs=None,
-)
+nwbfile = ecephys_session_to_nwb(session_key=session_key,
+                                 raw=True,
+                                 spikes=True,
+                                 lfp="dj",
+                                 end_frame=100,
+                                 lab_key=lab_key,
+                                 project_key=project_key,
+                                 protocol_key=protocol_key,
+                                 nwbfile_kwargs=None)
 
 nwbfile
 
@@ -101,7 +98,7 @@ nwbfile
 
 # +
 import time
-
+    
 write_nwb(nwbfile, f'./temp_nwb/{time.strftime("_test_%Y%m%d-%H%M%S.nwb")}')
 # -
 
@@ -116,16 +113,17 @@ write_nwb(nwbfile, f'./temp_nwb/{time.strftime("_test_%Y%m%d-%H%M%S.nwb")}')
 #
 # These values can be added to your `dj.config` as follows:
 
-dj.config["custom"]["dandiset_id"] = "<six digits as string>"
-dj.config["custom"]["dandi.api"] = "<40-character alphanumeric string>"
+dj.config['custom']['dandiset_id']="<six digits as string>" 
+dj.config['custom']['dandi.api']="<40-character alphanumeric string>"
 
 # This would facilitate routine updating of your dandiset.
 
 upload_to_dandi(
     data_directory="./temp_nwb/",
-    dandiset_id=dj.config["custom"]["dandiset_id"],
+    dandiset_id=dj.config['custom']['dandiset_id'],
     staging=True,
     working_directory="./temp_nwb/",
-    api_key=dj.config["custom"]["dandi.api"],
-    sync=False,
-)
+    api_key=dj.config['custom']['dandi.api'],
+    sync=False)
+
+